chore: ioredis only supports node
