+++ packages/remix-cloudflare/index.ts
@@ -21,7 +21,6 @@ export {
   createCookie,
   createCookieSessionStorage,
   createMemorySessionStorage,
-  createRedisSessionStorage,
   createSessionStorage,
 } from "./implementations";
