+++ packages/remix-deno/implementations.ts
@@ -2,7 +2,6 @@ import {
   createCookieFactory,
   createCookieSessionStorageFactory,
   createMemorySessionStorageFactory,
-  createRedisSessionStorageFactory,
   createSessionStorageFactory,
 } from "@remix-run/server-runtime";
 
@@ -14,5 +13,3 @@ export const createCookieSessionStorage =
 export const createSessionStorage = createSessionStorageFactory(createCookie);
 export const createMemorySessionStorage =
   createMemorySessionStorageFactory(createSessionStorage);
-export const createRedisSessionStorage =
-  createRedisSessionStorageFactory(createSessionStorage);
