+++ packages/remix-deno/index.ts
@@ -10,7 +10,6 @@ export {
   createCookie,
   createCookieSessionStorage,
   createMemorySessionStorage,
-  createRedisSessionStorage,
   createSessionStorage,
 } from "./implementations.ts";
