+++ packages/remix-node/package.json
@@ -27,7 +27,16 @@
   },
   "devDependencies": {
     "@types/cookie-signature": "^1.0.3",
-    "@types/source-map-support": "^0.5.4"
+    "@types/source-map-support": "^0.5.4",
+    "ioredis": "^5.3.1"
+  },
+  "peerDependencies": {
+    "ioredis": "^5.3.1"
+  },
+  "peerDependenciesMeta": {
+    "ioredis": {
+      "optional": true
+    }
   },
   "engines": {
     "node": ">=14"
