+++ packages/remix-server-runtime/interface.ts
@@ -8,4 +8,3 @@ export type {
 } from "./sessions";
 export type { CreateCookieSessionStorageFunction } from "./sessions/cookieStorage";
 export type { CreateMemorySessionStorageFunction } from "./sessions/memoryStorage";
-export type { CreateRedisSessionStorageFunction } from "./sessions/redisStorage";
