+++ packages/remix-server-runtime/package.json
@@ -26,16 +26,7 @@
   },
   "devDependencies": {
     "@remix-run/web-file": "^3.0.2",
-    "@types/set-cookie-parser": "^2.4.1",
-    "ioredis": "^5.3.1"
-  },
-  "peerDependencies": {
-    "ioredis": "^5.3.1"
-  },
-  "peerDependenciesMeta": {
-    "ioredis": {
-      "optional": true
-    }
+    "@types/set-cookie-parser": "^2.4.1"
   },
   "engines": {
     "node": ">=14"
